v0.0.11

+++ server/internal/usecase/process.go
@@ -20,8 +20,8 @@ type UploadFile struct {
 }
 
 type Point struct {
-	X float32 `json:"x"`
-	Y float32 `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 func NewProcessUsecase(ies domain.IImageEditService, iss domain.IImageStorageService, sss domain.ISessionStoreService) *ProcessUsecase {
